+++ advanced_alchemy/service/typing.py
@@ -6,34 +6,59 @@ should be a SQLAlchemy model.
 
 from __future__ import annotations
 
+from functools import lru_cache
 from typing import (
+    TYPE_CHECKING,
     Any,
+    Dict,
+    List,
+    Sequence,
+    TypeVar,
+    Union,
     cast,
 )
 
-from typing_extensions import TypeGuard
+from typing_extensions import Annotated, TypeAlias, TypeGuard
 
+from advanced_alchemy.repository.typing import ModelT
 from advanced_alchemy.service._typing import (
     LITESTAR_INSTALLED,
     MSGSPEC_INSTALLED,
     PYDANTIC_INSTALLED,
-    PYDANTIC_USE_FAILFAST,
     UNSET,
     BaseModel,
-    BulkModelDictT,
     DTOData,
     FailFast,
-    FilterTypeT,
-    ModelDictListT,
-    ModelDictT,
-    ModelDTOT,
-    ModelT,
-    PydanticOrMsgspecT,
     Struct,
     TypeAdapter,
     convert,
-    get_type_adapter,
 )
+
+if TYPE_CHECKING:
+    from advanced_alchemy.filters import StatementFilter
+
+PYDANTIC_USE_FAILFAST = False  # leave permanently disabled for now
+
+
+T = TypeVar("T")
+
+
+FilterTypeT = TypeVar("FilterTypeT", bound="StatementFilter")
+ModelDTOT = TypeVar("ModelDTOT", bound="Struct | BaseModel")
+PydanticOrMsgspecT = Union[Struct, BaseModel]
+ModelDictT: TypeAlias = Union[Dict[str, Any], ModelT, Struct, BaseModel, DTOData[ModelT]]
+ModelDictListT: TypeAlias = Sequence[Union[Dict[str, Any], ModelT, Struct, BaseModel]]
+BulkModelDictT: TypeAlias = Union[Sequence[Union[Dict[str, Any], ModelT, Struct, BaseModel]], DTOData[List[ModelT]]]
+
+
+@lru_cache(typed=True)
+def get_type_adapter(f: type[T]) -> TypeAdapter[T]:
+    """Caches and returns a pydantic type adapter"""
+    if PYDANTIC_USE_FAILFAST:
+        return TypeAdapter(
+            Annotated[f, FailFast()],
+        )
+    return TypeAdapter(f)
 
 
 def is_dto_data(v: Any) -> TypeGuard[DTOData[Any]]:
@@ -112,6 +137,7 @@ __all__ = (
     "Struct",
     "convert",
     "UNSET",
+    "UnsetType",
     "is_dto_data",
     "is_dict",
     "is_dict_with_field",
@@ -124,3 +150,19 @@ __all__ = (
     "is_pydantic_model_without_field",
     "schema_dump",
 )
+
+if TYPE_CHECKING:
+    if not PYDANTIC_INSTALLED:
+        from advanced_alchemy.service._typing import BaseModel, FailFast, TypeAdapter
+    else:
+        from pydantic import BaseModel, FailFast, TypeAdapter  # type: ignore[assignment] # noqa: TCH004
+
+    if not MSGSPEC_INSTALLED:
+        from advanced_alchemy.service._typing import UNSET, Struct, UnsetType, convert
+    else:
+        from msgspec import UNSET, Struct, UnsetType, convert  # type: ignore[assignment]  # noqa: TCH004
+
+    if not LITESTAR_INSTALLED:
+        from advanced_alchemy.service._typing import DTOData
+    else:
+        from litestar.dto import DTOData  # type: ignore[assignment] # noqa: TCH004
